feat: Better error handling and checks to prevent unnecessary deletes if network endpoints malfunction

+++ src/services/scheduler/network/updater.ts
@@ -3,6 +3,7 @@ import cluster from "cluster";
 import { KubeConfigService } from "src/config/kube/configuration.service";
 import { KubeService } from "src/kube/kube.service";
 import { NetworkService } from "src/network/network.service";
+import { SitesService } from "src/sites/sites.service";
 
 interface NetworkIngress {
   siteName: string;
@@ -23,21 +24,42 @@ export class NetworkUpdater {
     private kube: KubeService,
     private kc: KubeConfigService,
     private network: NetworkService,
+    private site: SitesService,
   ) {}
 
   async run(n) {
     this.logger.debug(`Updating network ${n}...`);
 
+    let clusterIngresses: ClusterIngress[];
+    let networkIngresses: NetworkIngress[];
+    let deleteIgressList: ClusterIngress[];
     // Get a list of all ingresses in the cluster.
-    const clusterIngresses = await this.kube.getIngresses(this.kc.namespace);
-    this.logger.debug(`clusterIngresses: ${JSON.stringify(clusterIngresses)}`);
+
+    try {
+      clusterIngresses = await this.kube.getIngresses(this.kc.namespace);
+      this.logger.debug(
+        `clusterIngresses: ${JSON.stringify(clusterIngresses)}`,
+      );
+    } catch (error) {
+      this.logger.error(`Error fetching cluster ingress list ${error}`);
+
+      throw error;
+    }
 
     // Get a list of all ingresses in the network.
-    const networkIngresses = await this.network.getIngressList();
-    this.logger.debug(`networkIngresses: ${JSON.stringify(networkIngresses)}`);
+    try {
+      networkIngresses = await this.network.getIngressList();
+      this.logger.debug(
+        `networkIngresses: ${JSON.stringify(networkIngresses)}`,
+      );
+    } catch (error) {
+      this.logger.error(`Error fetching network ingress list: ${error}`);
+
+      throw error;
+    }
 
     // Compare the clusterIngresses to networkIngresses and
-    // find items that are not in the cluster.
+    // find ingresses that are not in the cluster.
     const createIngressList: NetworkIngress[] = networkIngresses.filter(
       (networkIngress) => {
         return !clusterIngresses.some(
@@ -50,18 +72,23 @@ export class NetworkUpdater {
       `ingresses to create: ${JSON.stringify(createIngressList)}`,
     );
 
-    // Delete ingresses that are not in the network.
-    const deleteIgressList: ClusterIngress[] = clusterIngresses.filter(
-      (clusterIngress) => {
+    // Find ingresses that should be deleted
+    if (networkIngresses.length > 0) {
+      deleteIgressList = clusterIngresses.filter((clusterIngress) => {
         return !networkIngresses.some(
           (networkIngress) => networkIngress.siteName === clusterIngress.site,
         );
-      },
-    );
-
-    this.logger.debug(
-      `ingresses to delete: ${JSON.stringify(deleteIgressList)}`,
-    );
+      });
+      this.logger.debug(
+        `ingresses to delete: ${JSON.stringify(deleteIgressList)}`,
+      );
+      this.logger.debug(
+        `length of ingresses to delete: ${deleteIgressList.length}`,
+      );
+    } else {
+      deleteIgressList = [];
+      this.logger.debug("No ingresses to delete");
+    }
 
     // Add ingresses that are not in the cluster.
     for (const i of createIngressList) {
@@ -69,26 +96,35 @@ export class NetworkUpdater {
       const sitename = i.siteName;
       const deploymentUrl = i.deploymentUrl;
       const domain = i.domain;
-      try {
-        await this.kube.createIngress(
-          this.kc.namespace,
-          sitename,
-          deploymentUrl,
-          domain,
-        );
-      } catch (error) {
-        this.logger.error(`Error creating ingress: ${error}`);
+
+      const hasAddress = await this.site.checkDnsARecord(domain);
+
+      if (hasAddress) {
+        try {
+          await this.kube.createIngress(
+            this.kc.namespace,
+            sitename,
+            deploymentUrl,
+            domain,
+          );
+        } catch (error) {
+          this.logger.error(`Error creating ingress: ${error}`);
+        }
+      } else {
+        this.logger.error(`No DNS A record found for ${domain}`);
       }
     }
 
     // Delete ingresses that should be deleted
-    for (const ingress of deleteIgressList) {
-      try {
-        const ingressName = ingress.name;
-        this.logger.debug(`Deleting ingress ${ingressName}`);
-        await this.kube.deleteIngress(ingressName, this.kc.namespace);
-      } catch (error) {
-        this.logger.error(`Error deleting ingress: ${error}`);
+    if (deleteIgressList.length > 0) {
+      for (const ingress of deleteIgressList) {
+        try {
+          const ingressName = ingress.name;
+          this.logger.debug(`Deleting ingress ${ingressName}`);
+          await this.kube.deleteIngress(ingressName, this.kc.namespace);
+        } catch (error) {
+          this.logger.error(`Error deleting ingress: ${error}`);
+        }
       }
     }
   }
